function for creating invoices and added some colour coding for easy distinctions

+++ internal/database/add_stock.sql.go
@@ -25,7 +25,7 @@ type CreateStockItemParams struct {
 	Stockname string
 	Amount    int32
 	QtyType   string
-	PricePer  int32
+	PricePer  float64
 }
 
 func (q *Queries) CreateStockItem(ctx context.Context, arg CreateStockItemParams) (Stock, error) {
